import: src/sbin/gpt with rename
Note: importing as fork named fgdisk
Why?
Cause gpt(8) is dead upstream (replaced by "geom" smth)
and since it deals with bootcodes, better just to play it safe.
Why fgdisk?
Cause gpt(8) to deal with GPT is confusing as feck!
(and I'm not getting into "compatibility slices", for now)
It gives a clear warning on usage, think fdisk for gpt.
Thus fRAgdisk it is.

imported at 3268486 (2016-02-16)