sbin/fgdisk: Adjust PMBR cyl constants as in NetBSD
This caught my eye while trying to make bios-gpt booting
on my HP workstation. Too many 0xff!

Have no idea of implications of this, but given that
PMBR purpose is to protect gpt disk from fdisk lets try it.

If it is good enough for NetBSD it is good enough for me.

Taken-from: NetBSD (2013-12-04 21:59:47 +0200)